R6 Class representing a DB schema/table

The fundamental way to communicate with database tables.

Details

This class is a representation of a database table. It is the way that you can access data (e.g. tbl()), manipulate data (e.g. insert_data, upsert_data), and manipulate structural aspects of the database table (e.g. add_indexes, drop_indexes).

Public fields

conn

Database connection.

db_config

Configuration details of the database.

table_name

Name of the table in the database.

table_name_fully_specified

Fully specified name of the table in the database (e.g. [db].[dbo].[table_name]).

field_types

The types of each column in the database table (INTEGER, DOUBLE, TEXT, BOOLEAN, DATE, DATETIME).

field_types_with_length

The same as field_types but with (100) added to the end of all TEXT fields.

keys

The combination of variables that uniquely identify each row in the database.

keys_with_length

The same as keys but with (100) added to the end of all TEXT fields.

indexes

A named list of vectors (generally "ind1", "ind2", etc.) that improves the speed of data retrieval operations on a database table.

validator_field_contents

A function that validates the data before it is inserted into the database.

info

Free text information about the DB schema.

load_folder

A temporary folder that is used to write data to before inserting into the database.

censors

A named list of censors.

load_folder_fn

A function that generates load_folder.

Active bindings

load_folder_fn

A function that generates load_folder.

Methods

Public methods

Schema_v8$new()

Schema_v8$load_folder_fn()

Schema_v8$print()

Schema_v8$cat_status()

Schema_v8$is_connected()

Schema_v8$connect()

Schema_v8$disconnect()

Schema_v8$create_table()

Schema_v8$drop_table()

Schema_v8$insert_data()

Schema_v8$upsert_data()

Schema_v8$drop_all_rows()

Schema_v8$drop_rows_where()

Schema_v8$keep_rows_where()

Schema_v8$drop_all_rows_and_then_upsert_data()

Schema_v8$drop_all_rows_and_then_insert_data()

Schema_v8$tbl()

Schema_v8$print_dplyr_select()

Schema_v8$get_config_last_updated()

Schema_v8$add_indexes()

Schema_v8$drop_indexes()

Schema_v8$db_connect()

Schema_v8$db_disconnect()

Schema_v8$clone()

Method new()

Create a new Schema_v8 object.

This function is not used directly. You should use add_schema_v8 instead.

Usage

Schema_v8$new(
  conn = NULL,
  db_config = NULL,
  table_name,
  field_types,
  keys,
  censors = NULL,
  indexes = NULL,
  validator_field_types = validator_field_types_blank,
  validator_field_contents = validator_field_contents_blank,
  info = NULL
)

Arguments

conn

Database connection (generally leave empty).

db_config

Configuration details of the database.

table_name

Name of the table in the database.

field_types

The types of each column in the database table (INTEGER, DOUBLE, TEXT, BOOLEAN, DATE, DATETIME).

censors

censors A named list of censors.

indexes

A named list of vectors (generally "ind1", "ind2", etc.) that improves the speed of data retrieval operations on a database table.

validator_field_types

A function that validates the field_types before the DB schema is created.

validator_field_contents

A function that validates the data before it is inserted into the database.

Returns

A new Schema_v8 object.

Method load_folder_fn()

Usage

Schema_v8$load_folder_fn()

Method print()

Class-specific print function.

Usage

Schema_v8$print(...)

Method cat_status()

Displays the status (connected/disconnected)

Usage

Schema_v8$cat_status()

Method is_connected()

Is the DB schema connected?

Usage

Schema_v8$is_connected()

Returns

TRUE/FALSE

Method connect()

Connect to the database

Usage

Schema_v8$connect(db_config = self$db_config)

Arguments

db_config

db_config

Method disconnect()

Disconnect from the database

Usage

Schema_v8$disconnect()

Method create_table()

Create the database table

Usage

Schema_v8$create_table()

Method drop_table()

Drop the database table

Usage

Schema_v8$drop_table()

Method insert_data()

Usage

Schema_v8$insert_data(newdata, verbose = TRUE)

Arguments

newdata

The data to insert.

verbose

Boolean. Inserts data into the database table

Method upsert_data()

Upserts data into the database table

Usage

Schema_v8$upsert_data(
  newdata,
  drop_indexes = names(self$indexes),
  verbose = TRUE
)

Arguments

newdata

The data to insert.

drop_indexes

A vector containing the indexes to be dropped before upserting (can increase performance).

verbose

Boolean.

Method drop_all_rows()

Drops all rows in the database table

Usage

Schema_v8$drop_all_rows()

Method drop_rows_where()

Drops rows in the database table according to the SQL condition.

Usage

Schema_v8$drop_rows_where(condition)

Arguments

condition

SQL text condition.

Method keep_rows_where()

Keeps rows in the database table according to the SQL condition.

Usage

Schema_v8$keep_rows_where(condition)

Arguments

condition

SQL text condition.

Method drop_all_rows_and_then_upsert_data()

Drops all rows in the database table and then upserts data.

Usage

Schema_v8$drop_all_rows_and_then_upsert_data(
  newdata,
  drop_indexes = names(self$indexes),
  verbose = TRUE
)

Arguments

newdata

The data to insert.

drop_indexes

A vector containing the indexes to be dropped before upserting (can increase performance).

verbose

Boolean.

Method drop_all_rows_and_then_insert_data()

Drops all rows in the database table and then inserts data.

Usage

Schema_v8$drop_all_rows_and_then_insert_data(newdata, verbose = TRUE)

Arguments

newdata

The data to insert.

verbose

Boolean.

Method tbl()

Provides access to the database table via dplyr::tbl.

Usage

Schema_v8$tbl()

Method print_dplyr_select()

Prints a template dplyr::select call that you can easily copy/paste for all your variables.

Usage

Schema_v8$print_dplyr_select()

Method get_config_last_updated()

Extracts the corresponding row of data from get_config_last_updated.

Usage

Schema_v8$get_config_last_updated()

Method add_indexes()

Adds indexes to the database table from self$indexes

Usage

Schema_v8$add_indexes()

Method drop_indexes()

Drops all indees from the database table

Usage

Schema_v8$drop_indexes()

Method db_connect()

Usage

Schema_v8$db_connect(db_config = self$db_config)

Method db_disconnect()

Usage

Schema_v8$db_disconnect()

Method clone()

The objects of this class are cloneable with this method.

Usage

Schema_v8$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.