task_from_config

task_from_config

task_from_config(
  name,
  type,
  cores = 1,
  db_table = NULL,
  filter = "",
  for_each_plan = NULL,
  for_each_argset = NULL,
  upsert_at_end_of_each_plan = FALSE,
  insert_at_end_of_each_plan = FALSE,
  action,
  schema = NULL,
  args = NULL
)

Arguments

name
Name of task

type
Type of task ("data","single", "analysis", "ui")

cores
Number of cores to run the task on

db_table
Database table

filter
Filter for the database table

for_each_plan
Create a plan for each value

for_each_argset
Create an argset for each value

upsert_at_end_of_each_plan
If TRUE, then schema$output is used to upsert at the end of each plan

insert_at_end_of_each_plan
If TRUE, then schema$output is used to insert at the end of each plan

action
Text

schema
List of schema mappings

args
List of args