Creating a task from config (v3)

This function is used to easily create a task

task_from_config_v3(
  name = NULL,
  name_grouping = NULL,
  name_action = NULL,
  name_variant = NULL,
  cores = 1,
  plan_argset_fn_name = NULL,
  for_each_plan = NULL,
  for_each_argset = NULL,
  universal_argset = NULL,
  upsert_at_end_of_each_plan = FALSE,
  insert_at_end_of_each_plan = FALSE,
  action_fn_name,
  data_selector_fn_name = NULL,
  schema = NULL,
  info = NULL
)

Arguments

name
Name of the task (dont use)

name_grouping
Name of the task (grouping)

name_action
Name of the task (action)

name_variant
Name of the task (variant)

cores
Number of CPU cores

plan_argset_fn_name
The name of a function that returns a named list list(for_each_plan = list(), for_each_argset = NULL).

for_each_plan
A list, where each unit corresponds to one data extraction. Generally recommended to use plnr::expand_list.

for_each_argset
A list, where each unit corresponds to one analysis within a plan (data extraction). Generally recommended to use plnr::expand_list.

universal_argset
A list, where these argsets are applied to all analyses univerally

upsert_at_end_of_each_plan
Do you want to upsert your results automatically at the end of each plan?

insert_at_end_of_each_plan
Do you want to insert your results automatically at the end of each plan?

action_fn_name
The name of the function that will be called for each analysis with arguments data, argset, schema

data_selector_fn_name
The name of a function that will be called to obtain the data for each analysis. The function must have the arguments argset, schema and must return a named list.

schema
A named list that maps sc::config$schemas for use in action_fn_name and data_selector_fn_name

info
Information for documentation